refactor: consolidate loader functions
