+++ launcher-boot/Makefile
@@ -5,7 +5,7 @@ STACKSIZE = 0x04000
 EE_BIN = payload.elf
 EE_BIN_STRIPPED = payload-stripped.elf
 EE_BIN_PACKED = payload-packed.elf
-EE_OBJS = main.o
+EE_OBJS = main.o ../common/loader.o
 EE_LDFLAGS += -Wl,--gc-sections -Wl,-Ttext -Wl,$(LOADADDR) -Wl,--defsym -Wl,_stack_size=$(STACKSIZE) -Wl,--defsym -Wl,_stack=$(STACKADDR)
 EE_LIBS += -lpatches -lfileXio
 EE_INCS += -I$(GSKIT)/include -I$(PS2SDK)/ports/include
